de-couple use-case-specific code from more generic code and incorporate Kenichi feedback
