+++ data-model/use_case_specific/pxrd_collector_mocof1.py
@@ -1,0 +1,3 @@
+def process_pxrd_file_use_case_specific(pxrd_file):
+    pxrd_file.sample_holder_shape = pxrd_file.sample_holder_shape.replace("film", "KAPTON_FILMS").replace("capillary",
+                                                                          "HILGENBERG_GLASS_NO_14_CAPILLARY")
